fix: make thumbnail uncontrolled component
